(PC-9905) simplify Jouve access verification in fraud_view
