+++ src/pcapi/admin/custom_views/fraud_view.py
@@ -116,7 +116,13 @@ class FraudView(base_configuration.BaseAdminView):
         return formatters
 
     def is_accessible(self) -> bool:
-        return flask_login.current_user.is_authenticated and flask_login.current_user.isAdmin
+        # TODO: remove when we have a clean way to get groups from google.
+        # This is a hackish way to filter from a user role which is weak : we do want a way
+        # to add permissions based on groups and sync'ed from our google IDP, and not developp a way to do it here.
+        if flask_login.current_user.is_authenticated and users_models.UserRole.JOUVE in flask_login.current_user.roles:
+            return True
+
+        return super().is_accessible()
 
     def get_query(self):
         filters = users_models.User.beneficiaryFraudChecks.any() | users_models.User.beneficiaryFraudResult.has()
