+++ tests/admin/custom_views/fraud_view_test.py
@@ -191,3 +191,35 @@ class JouveUpdateIDPieceNumberTest:
         fraud_factories.BeneficiaryFraudCheckFactory(type=fraud_models.FraudCheckType.DMS)
         response = client.get("/pc/back-office/beneficiary_fraud")
         assert "<ul><li>dms</li></ul>" not in response.data.decode()
+
+
+@pytest.mark.usefixtures("db_session")
+class JouveAccessTest:
+    """Specific tests to ensure JOUVE does not access anything else"""
+
+    def test_access_index(self, client):
+        user = users_factories.UserFactory(isAdmin=False, roles=[users_models.UserRole.JOUVE])
+        client.with_auth(user.email)
+        response = client.get("/pc/back-office/")
+        assert response.status_code == 200
+
+    def test_access_fraud_views(self, client):
+        user = users_factories.UserFactory(isAdmin=False, roles=[users_models.UserRole.JOUVE])
+        client.with_auth(user.email)
+        response = client.get("/pc/back-office/beneficiary_fraud")
+        assert response.status_code == 200
+
+    @pytest.mark.parametrize(
+        "url",
+        [
+            "/pc/back-office/pro_users",
+            "/pc/back-office/admin_users",
+            "/pc/back-office/beneficiary_users",
+        ],
+    )
+    def test_access_forbidden_views(self, client, url):
+        user = users_factories.UserFactory(isAdmin=False, roles=[users_models.UserRole.JOUVE])
+        client.with_auth(user.email)
+        response = client.get(url)
+        assert response.status_code == 302
+        assert response.headers["Location"] == "http://localhost/pc/back-office/"
